Rename "program" -> "programme". Minor design-updates to re-design
