+++ packages/designmanual/dummydata/mockPrograms.js
@@ -495,7 +495,7 @@ export const programs = [
 
 export const subjectCategories = [
   {
-    name: 'fellesfag',
+    name: 'Fellesfag',
     subjects: [
       {
         nodeId: '42',
@@ -550,7 +550,7 @@ export const subjectCategories = [
     ],
   },
   {
-    name: 'yrkesfag',
+    name: 'Yrkesfag',
     subjects: [
       {
         nodeId: '51',
@@ -641,7 +641,7 @@ export const subjectCategories = [
     ],
   },
   {
-    name: 'studieforberedende',
+    name: 'Studieforberedende',
     subjects: [
       { nodeId: '52234', name: 'Biologi 1', url: '#', id: 'urn:subject:42' },
       {
