+++ packages/ndla-ui/src/Navigation/NavigationBox.tsx
@@ -55,7 +55,7 @@ export type ItemProps = {
 };
 type Props = {
   heading?: string;
-  colorMode?: 'dark' | 'light';
+  colorMode?: 'dark' | 'light' | 'lighterGrey';
   items: [ItemProps];
   onClick?: (event: React.MouseEvent<HTMLElement>) => void;
 };
@@ -78,6 +78,7 @@ export const NavigationBox = ({
           <SafeLinkButton
             to={item.url}
             lighter={colorMode === 'light'}
+            lighterGrey={colorMode === 'lighterGrey'}
             darker={item.selected}
             buttonSize="medium"
             borderShape="sharpened"
