+++ packages/ndla-ui/src/index-javascript.js
@@ -174,4 +174,4 @@ export {
   NavigationBox,
   NavigationTopicAbout,
 } from './Navigation';
-export { default as Program } from './Program';
+export { default as Programme } from './Programme';
